Add class JPetOptionTypeHandler
This class is used to handle the type operation of user defined options.
