+++ JPetOptionsGenerator/JPetOptionsGenerator.h
@@ -46,11 +46,8 @@ public:
 
   bool isOptionSet(const std::map<std::string, boost::any>& optionsMap, const std::string& option) const;
   boost::any getOptionValue(const std::map<std::string, boost::any>& optionsMap, std::string option) const;
-  std::string getTypeOfOption(const std::string nameOfOption) const;
-  std::string getNameOfOption(const std::string option) const;
 
   std::map<std::string, boost::any> variablesMapToOption(const po::variables_map& variablesMap) const;
-  std::map<std::string, std::string> anyMapToStringMap(const std::map<std::string, boost::any>& optionsMap) const;
   std::map<std::string, std::vector<bool(*)(std::pair <std::string, boost::any>)> > generateValidationMap() const;
   std::map<std::string, std::vector<std::pair <std::string, boost::any>(*)(boost::any)> > generateTransformationMap() const;
 
@@ -71,14 +68,5 @@ public:
   static std::pair <std::string, boost::any>getLowerEventBound(boost::any option);
   static std::pair <std::string, boost::any>getHigherEventBound(boost::any option);
 
-private:
-  enum optionTypes {
-    Int,
-    String,
-    Bool,
-    VectorString,
-    VectorInt,
-    Default
-  };
 };
 #endif
